+++ CHANGELOG.md
@@ -22,14 +22,15 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - Implement content update using Json Patch operations [(#362)](https://github.com/wazuh/wazuh-indexer-plugins/pull/362)
 - Implement CTI API client rate limit and enhanced response handling [(#363)](https://github.com/wazuh/wazuh-indexer-plugins/pull/363)
 - Add custom action to list modified plugins [(#388)](https://github.com/wazuh/wazuh-indexer-plugins/pull/388)
-- Expose settings for the Content Manager plugin [#395](https://github.com/wazuh/wazuh-indexer-plugins/pull/395)
+- Expose settings for the Content Manager plugin [(#395)](https://github.com/wazuh/wazuh-indexer-plugins/pull/395)
+- Add content update to to the Content Manager init process [(#421)](https://github.com/wazuh/wazuh-indexer-plugins/pull/421)
 
 ### Dependencies
 -
 
 ### Changed
-- Update http components to latest versions [#403](https://github.com/wazuh/wazuh-indexer-plugins/pull/403)
-- Content Manager tier 1 final wrap up [#373](https://github.com/wazuh/wazuh-indexer-plugins/pull/373)
+- Update http components to latest versions [(#403)](https://github.com/wazuh/wazuh-indexer-plugins/pull/403)
+- Content Manager tier 1 final wrap up [(#373)](https://github.com/wazuh/wazuh-indexer-plugins/pull/373)
 - Refactor Content Manager's code and fix Catalog info indexing [(#317)](https://github.com/wazuh/wazuh-indexer-plugins/pull/317)
 - Improved mdbook installation instructions [#332](https://github.com/wazuh/wazuh-indexer-plugins/pull/332)
 - Third-party integrations maintenance [(#299)](https://github.com/wazuh/wazuh-indexer-plugins/pull/299) [(#374)](https://github.com/wazuh/wazuh-indexer-plugins/pull/374) [(#398)](https://github.com/wazuh/wazuh-indexer-plugins/pull/398)